Implemented JS channel-mapper function
Put the JavaScript source code in the doc _design/channels, in the property "channelmap".
It should call the function sync() to add the document to a channel.
Arguments to sync are one or more channel names (or arrays of names.)
The function can call sync zero or more times, just as with the regular 'emit'.

Changing the channelmap function in the design doc has NO effect until the server restarts.
After restart it will take effect for any documents subsequently updated, but not for existing
revisions of documents. (This will be improved in the future.)