REST interface for accessing user login credentials
Runs on a separate port since it needs to be blocked from public access.
Port defaults to :4985 but can be set via the -authaddr flag.
API looks like:

GET, PUT, DELETE /username --> Access the credentials for user 'username' as JSON:
{"name": "____", "password": "____", "channels": ["____", "____", …]}
POST / --> Creates a new user (name given in the JSON "name" property)

The name for the default user (if no Basic auth given) is "GUEST".
By default it has access to anything but you can use PUT to limit it to some channels,
or block it entirely by setting the channels to the empty array.

The channel name "*" is a wildcard and allows access to all documents
(even those that aren't in any channels!)