Speed up parameter passing to JS sync function
Use new walrus API that lets us pass Go values as input to a JS function
without having to marshal them to JSON first. This avoids marshaling
the new document's body and the userCtx in the call to the sync fn.

This requires the latest version of walrus, so run
go get -u github.com/couchbaselabs/walrus