Use "changes" view to fill in if the channel log is incomplete
If the channel log doesn't go back as far as the 'since' query param,
ChangesFeed() now runs a "changes" view query to get the earlier revs.

If the channel log is entirely missing, ChangesFeed() also rebuilds it
from the view output and saves the channel log. So it should be possible
to delete channel-log documents by hand to fix any corruption.