Use compound sequence IDs
Instead of sequence numbers, the _changes feed now uses compound IDs
that look like chan:##,chan:##… Basically, for each channel it lists
the last sequence number seen on that channel. (This isn't the same as
the _highest_ sequence number seen, since sequences might not be added
to a channel log in order. But knowing the last number seen tells the
changes feed where to start in the log next time.)