Public sequence IDs contain global last-sequence value
As an optimization, sequence IDs now contain an entry for channel "*"
whose value is (an approximation of) the database's last sequence number
overall. This allows the _changes feed implementation to quickly decide
there are no new changes.

This number appears in the sequence ID as the value of the channel "*".

Unfortunately it's tricky to determine the actual last sequence number
without running into race conditions, so the current implementation only
approximates it as the largest sequence number on a subscribed channel.
Hopefully this can be improved later on.