Use new Bucket.WriteUpdate to fix race condition with access()
If a doc revision changes user access to a channel, the updateDoc
method saves the new access in the doc metadata and then updates the
user doc to clear the cached channel-access map.
The problem is, the next time the User object is loaded and it rebuilds
the access map, the document revision that triggered the change might
not be saved to disk yet, so the view results will be stale, causing
obsolete channel access to be stored.
This patch fixes this by using the Indexable mode of the new UpdateWrite
call to ensure that the new doc revision is indexable.