Added admin REST API to perform CPU profiling
A POST to /_profile on the admin port turns CPU profiling on/off.
To turn on profiling, add a JSON "file" property whose value is the file
path to write the profile to.
To turn off profiling, omit the "file" property.

The profile file can be analyzed using the Go 'pprof' tool -- see
http://blog.golang.org/profiling-go-programs