Updating version (and commit hash) on each build
With binaries, it can be hard to keep track of which version or specific commit you may be running (especially when trying to switch between versions a lot, using multiple machines, etc). The enhancements to build.sh grab branch and commit info, and then they're displayed alongside the VersionString to keep you up-to-date.