Fixed incorrect error on GET of unknown database
It was trying to contact the config server even if none was configured,
resulting in a POST to a URL of "", and then returned that error.
Fixed that, and made config-server-related errors more distinctive.
Oh, and while I was in there I had it return a 400 on an illegal db name.
Fixes #171