Fixed nasty bug with change-log generation that caused duplicate entries
addChangeLog() -- which saves a new channel log if one doesn't already
exist -- was not running on the same goroutine as the channelLogWriter
that makes all other changes to that channel log. That meant it could
add the channel before the entries got written, and then the entries
would be appended to that channel as duplicates.