Limit number of incoming HTTP connections
To prevent the gateway from getting clogged up with zillions of HTTP
connections that are mostly blocked waiting to talk to the database, I
added a layer that limits the number of simultaneously open HTTP
sockets.
The default is 1000, but you can set the top-level config property
"maxIncomingConnections" to customize it. A zero value turns off the
limits entirely.
Also, I renamed the recently-added limits on Couchbase database
connections to "maxCouchbaseConnections" and "maxCouchbaseOverflow".