Added WebSocket support for the _changes feed
Use “?feed=websocket” to enable it. It’s comparable to continuous feed
except it upgrades the connection to WebSocket and sends groups of
changes as messages. Each message is a JSON array of change
dictionaries. An empty array means the server is currently waiting
for new changes; this can be used by the client to detect when
it’s “caught up”.
Fixes #143