Generalized profiling support
You can now capture any registered Go profile “profilename” by:
    POST /_profile/profilename
See the Go docs for pprof.Profile to see what profiles are available.
Standard ones are “heap”, “goroutine”, “threadcreate” and “block”.
The body of the request needs to be JSON with a “file” property
containing the path to write the profile to.