Admin API: Extensions to _all_docs to get channel and access info
If you get /db/_all_docs?channels=true from the admin port, each row’s
“value” will include a “channels” property whose value is an array of
the channels that document is in.

With ?access=true, there’s an “access” property whose value is a
dictionary mapping user and role names to arrays of channels they’re
granted access to by this document. (Role names are prefixed with
“role:”.)

If you want to get info for only specific documents, you can either use
the startkey and endkey parameters, or POST with a JSON body that’s an
array of the document ID(s).