Admin ui included as Go code
The admin ui is available on the admin port at the /_utils/ path.
It is delivered as a single HTML asset which is copied from the
utils/ submodule into the sync_gateway rest package before build.

There is a config parameter (adminUI) that tells the server to look on the
filesystem for HTML to serve instead of the packaged version. This
supports both rapid development and customization.