Better error handling in _bulk_get
After we start writing the response we can’t return an HTTP error status
anymore. So any error that occurs while processing one of the requested
docs in the _bulk_get request has to be returned as an error value in
the multipart response, not by returning an error from the handler
method.