Move adminUI to vendor. Much simpler.
The admin ui is available on the admin port at the /_admin/ path.
It is delivered as a single HTML asset which is linked from the
adminUI/ submodule into the sync_gateway GOPATH in go.sh and test.sh.

There is a config parameter (adminUI) that tells the server to look on the
filesystem for HTML to serve instead of the packaged version. This
supports both rapid development and customization / independent upgrade.