Show conflicting docs in _changes feed
An entry for a doc in conflict should contain multiple _revs entries,
so that a server-side app reading the feed can tell there's a conflict.
Note that the feed might still contain multiple entries for the same
doc in some situations; but now the later one will definitely show that
there's a conflict.
Fixes #242