Give clear & obvious warning if trying to build with older Go compiler
This should help people trying to build the Sync Gateway when they don't
have Go 1.2; I've gotten two reports of this in the past day alone.

Note that due to my very limited bash fu, this will also fail for people
using a _newer_ version of Go like a dev build of 1.3. To work around
this you'll either need to locally hack your go.sh script, or fix it
to do a smarter job of version checking.