Changes feeds now correctly handle role changes
If a user is granted a new role, and that role grants access to new
channels, then the next call to the changes feed should return all
sequences in those newly-added channels.
This required changing the structure of a user document to store the
dynamic channel list as a TimedSet instead of an array of strings, so
the user knows when it was added to each role.