Properly invalidate users/roles when doc is deleted
A deleted doc doesn't grant access to any channels or roles. But deleting
a doc didn't invalidate the affected users/roles' caches, so they
wouldn't always update correctly. Fixed this by having the deletion
always set the doc's channel/role access to empty sets.

This doesn't actually fix #337, but it reduces it to the already known
issue that losing access to a channel doesn't causing its docs to be
revoked in the changes feed (#264).