Fixed performance regression (and incorrect _changes feeds)
Regression was in commit da9037da. options.Since wasn't being updated
during the MultiChangesFeed inner loop, so in a continuous changes feed,
every time the database changed it would re-emit all the changes since
the original Since parameter, not just the new ones. This created a
huge amount of unnecessary output, consuming a lot of CPU and I/O and
causing the gateway to fall behind in processing data.
Fixes #406