Added admin API for design docs
The admin REST API now supports the endpoint /db/$ddoc/_view/$name for
design docs.

For now this is just a passthru to Couchbase Server's API,
but the plan is to modify the map functions to (a) skip special docs,
(b) strip out the "_sync" metadata, and (c) record info about the
channels of each row's doc so we can apply filtering at query time.
For #379