Switch full event queue handling to drop incoming instead of oldest
Event queue was dropping the oldest in the queue when the queue was full - changed to instead block incoming when queue full. Changed the raiseEvent method to return err on full queue, to allow consumers to handle if appropriate.  Currently not used for document change event - expected as future handling for failed events.

Also includes a couple of minor code tweaks based on review, and some refactoring for code readability.