Fix for race condition when creating database config based on ConfigServer
Modified the mutex handling in server_context.go to handle the race condition.  Renamed AddDatabaseFromConfig to 'getOrAddDatabaseFromConfig'.  It now gets a write lock for the server context mutex, and then does a read to see if another process has already updated the config.  It uses a new 'useExisting' parameter, to differentiate between the cases where we want to reuse an existing connection (the ConfigServer scenario), or throw an error on duplicate (all other scenarios).  The original AddDatabaseFromConfig now just calls getOrAddDatabaseFromConfig with useExisting=false.

To avoid future confusion about usage of the mutex lock, moved a couple of methods that were only getting used by AddDatabaseFromConfig inline into getOrAddDatabaseFromConfig.