Refactor late feed handling to reduce performance overhead
Removed the read lock on late sequence processing by the changes feed.  The potential for duplicate late sequences to be sent on the changes feed should be low, and would be handled by existing RevsDiff processing on the client.  Duplicates only arrive in the scenario where the user has visibility to a late arriving sequence via multiple channels, and the _changes processing finds it for each channel in different iterations of the processing loop.  Once performance is tested on the branch, can review whether alternate handling (comparing late sequences found to those served in previous iterations) is required.

Streamlined the check to see whether new late sequences have arrived, and corrected purge handling for the last sequence queue.

Removed unnecessary logging on the main _changes loop.