Added delta-compression support for getting JSON docs
When deltas are enabled in a request, the JSON body may now also be
delta-compressed. The source of the delta will be the first available
revision listed in the `atts_since` parameter. This revision ID will
appear in the `X-Delta-Source` header.

See https://github.com/couchbaselabs/couchbase-lite-api/wiki/Delta-Compression
For #452; see also couchbase/couchbase-lite-ios#168