Use Go's sync.Pool to maintain cache of gzip.Writers
This is conceptually similar to the existing approach that uses a
channel as a cache; but the sync.Pool class is tied into the Go runtime
and GC, so it should be able to make better decisions about memory
management.

I made the gzip.Writer cache available via functions so that other code
in SG that uses gzip can take advantage of it. (Right now that code is
down in db/attachment.go, but in my delta branch it's moving into the
rest package.)

Also updated version requirement in go.sh, because sync.Pool is a Go 1.3
feature.