GZip-compress attachments that don't have delta sources
If there isn't an earlier version of an attachment available (e.g. it's
a new attachment), so it can't be delta-compressed, use regular GZip
instead, assuming the metadata indicates it's compressible.

Fixed an ambiguity in the Attachment struct: the `deltaSource` field
indicated both a _potential_ delta source and the _actual_ delta source,
so there wasn't a way to tell whether the attachment data is a delta.