Delta-compress encoded attachments
Previously, attachments with "encoding":"gzip" weren't eligible for
delta compression because their data is gzipped which obviously won't
delta-compress well. But gzip-encoded attachments tend to be those that
are very compressible inside the encoding (text and such), so it's worth
delta-encoding them. We just have to decode the gzip first.