Avoid errors when user is deleted during changes processing
When a user with an active continuous or longpoll _changes feed is deleted through the admin API, the changes feed will now be closed.  This was previously causing a panic when the user record was reloaded during changes processing.  Closing of feed is communicated back to the handler by a new error property on the ChangesEntry.

Fixes #809.