Restore support for querying reduce views through the admin API
Fixes #857

Restores the ability to query views with a reduce function when running queries that don't require channel filtering.  There are two scenarios where channel filtering isn't applied (and so reduce functions can be used):
1. When running the query over the admin API.
2. When the user running the query has access to the `*` channel

If a view with a reduce function is used in any other scenario, it returns no data and logs a warning in the Sync Gateway logs.