Support for max heartbeat parameter in SG config
Allows users to define a maximum heartbeat value for continuous changes feeds for a Sync Gateway node.  Primary use case is when Sync Gateway is deployed behind a proxy/load balancer with a known timeout - it's more efficient to define the max heartbeat on Sync Gateway, instead of depending on clients to scale back their heartbeat in response to closed connections.

Required refactoring of the existing getRestrictedInt... handling to address a couple of issues:
1. Allow use of '0' to mean 'disabled'.  e.g. clients explicitly sending heartbeat=0 will result in no heartbeats being sent.  Previously incoming requests were converting 0 to minValue.
2. Continue to support 'omitted' as 'default'.  e.g. clients not sending a heartbeat value will use the default heartbeat value.
3. Make readChangesOptionsFromJSON a handler method, to enable access to the new heartbeat value from config.

Unit tests have good coverage of JSON handling (via TestReadChangesOptionsFromJSON), but there isn't a good way to exercise the GET _changes flow via the current unit tests - there's no easy way to get visibility into the heartbeat/timeout values after making an httptest invocation of _changes.  Tested this flow in standalone SG to ensure it works as expected.