Materialized reduce queries
This allows Sync Gateway to convert reduce queries into documents
that can be synced on channels. So eg each user can sync their
activity to Sync Gateway, and a global activity counter will sync
back down with the aggregate results.

The API is an "into" parameter on view queries, which copies them
into a target database. If into=true the target database is the
same as the source database. Otherwise, into can be a database name.