Add LRU cache to sequence hasher
Stores up to 500 recently used hash values in an LRU cache.  Implementation is based on revision cache, with modifications to support changes to the set of clocks for a hash within the cache (hash collisions).  Reduces the number of bucket ops required when multiple connected users have the same set of channels.  On cache hit, uses the cache value.  On cache miss, writes to the hash bucket then updates the cache.  Existing CAS write processing handles the case where another Sync Gateway node has already written that value to the bucket - cached value will be updated with the current state of the hash bucket value.

Two-level locking, similar to revision cache.  The entire cache only gets locked during read/write of the cache entry.  A lock on the individual entry is used when doing bucket operations, to avoid blocking for other cache readers.