For Amazon Linux, will set OS to RedHat and will default versions to 5.0.
The Amazon Linux distort removes the /etc/redhat-release file and there is no direct way to determine the REDHAT/CentOS release it is base on. Amazon say that it is binary compatible with at least RedHat 5, later releases are binary compatible with RedHat 6. Defaulting to RedHat 5 as the sys_init scripts should be compatible with both.