Improve handling for non-indexed channels
Previously a request for changes for a channel not found in the index would return an empty set, but not create a channel reader.  This resulted in a high number of queries to that channel clock, as the caching and polling usually done by the channel index reader weren't kicking in.