To support running without a global sequence (e.g. using distributed index), user and role access handling needs to be enhanced to support vbucket sequences.
TimedSet has been refactored to support a collection of VbSequences (vbucket,sequence pairs).  When the vbucket is not specified, TimedSets work as previously, including JSON marshalling into the previous format.  When vbucket is specified, TimedSets are marshalled to include both vb and seq for each channel.

VbSequence handling for ExplicitChannels (admin-granted channels) and access-based grants are included.  Removes previous handling to duplicate the user channel information in the index bucket - for performance, this needs to be incorporated into the user document itself.

Backfill handling for _changes requests (in index_changes.go) is enhanced to support using a sequence clock/hash as a triggered by value, and to support resuming partial backfill without a global sequence.