Avoid looping on one-off changes requests with low sequence
When a one-off changes request comes in with a low sequence included (e.g. since="2::5"), we want to use the same handling we're already doing for longpoll - ignore the low sequence value if it matches the SG's current lowSequence.  This avoids infinite looping of the records between low::high.  It also means any additional skipped sequences between low::high won't be sent until low arrives or is abandoned.