Avoid race between SG nodes when writing index partition map
When a new channel index is initialized, Sync Gateway writers will initialize the partition map based on information from CBGT when it's not found in the index bucket, and then persist the CBGT information to the index bucket.  There were two problems happening:

1. Sync Gateway nodes were doing a Set when they persisted the partition map, instead of an Add.  This was the race - two sync gateways could look for the partition map in the DB, not find it, then write new partition maps to the based on what they got from CBGT.
2. The ordering of partitions in the slice returned by CBGT isn't constant - there's an underlying map range happening.  As a result, Sync Gateways hitting this race condition would end up using two different partition maps (and clobbering each others partition map writes to the bucket)

The fix covers both aspects:
1. The partitions in the slice returned by CBGT are sorted by UUID before any processing by Sync Gateway.
2. Sync Gateway does an add instead of set when writing the partition map to the bucket, and when the key already exists, reloads and uses the version from the bucket.

Strictly speaking only one of these fixes are necessary for this issue, but adding both is better future-proofing for changes for either.