Ensure all sequence values returned by changes feed can be used as since value
When using vector clock sequences (distributed index), the changes feed returns a hashed sequence clock, instead of an integer sequence.  Previously longpoll and one-off changes requests were only hashing the last_seq value, to avoid the performance overhead of hashing a sequence clock for each changes entry - other entries were sent as simple vb.seq format.  However, this breaks clients (including CBL) that attempt to make a subsequent _changes request based on a change entry's sequence value, instead of last_seq.

To handle this, longpoll and one-off changes now use the same approach as previously used for continuous changes feeds - the cumulative sequence clock for a changes response is hashed periodically, and the most recent hash is sent as a `low` value for each changes entry (e.g. in the format `low::vb.seq`, where low is a clock hash.

To ensure that the last sequence sent in one-off and longpoll replication gets a new hash value (to avoid looping of longpoll requests), the changes loop processing was refactored in index_changes to identify the last sequence before it is sent on the output channel.

Included in this PR is the ability to define the frequency of the periodic sequence hashing in the Sync Gateway config, as an optional part of the channel_index configuration.  This also exposes the other sequence hash options that weren't previously exposed, including the ability to define a separate bucket for sequence hashes, and to customize the expiry for persisted hash values.  For 1.2 this seq_hashing config will be released as an undocumented feature, with the plan for full QE testing and documentation in 1.3.