Principal notification when using distributed index
Changes to principal docs (users, roles) weren't triggering notification when Sync Gateway was running as an index reader in distributed channel index mode, with the exception of single node reader-writers.

Updates index writers to increment principal counts as part of processPrincipalDoc, and updates index readers to poll for changes to principal counts in the same way that's done for channels.

In 1.3, targeted for review to move away from a polling approach, along with the channel handling, for the GA release of distributed index.

Fixes #1524.