Disable watchDocChanges listening unless needed
watchDocChanges listens to an echoed channel that gets populated with non-system docs that are seen on the server mutation feed, and is used to drive autoImport or bucket shadowing.   It's now disabled unless the database is actually configured for autoImport or bucket shadowing.

The channel used for watchDocChanges is initialized in NewDatabaseContext, which is called before the shadower is initialized (when present), so we can't just check for shadower != nil at that point.  To pass this information into NewDatabaseContext, added a new TrackDocs property to the DatabaseContextOptions to pass the value in.