Improve attachment revpos handling for pruned ancestors
Document updates including previously existing attachments were failing when an ancestor revision body is no longer available, as SG isn't able to match the incoming revpos with an attachment on non-compacted ancestor body.  The most common scenario is the addition of a conflicting revision under a revision that is no longer in the rev cache, and the temporary revision copy has expired out of the DB.

This commit adds an additional check in this scenario: if the incoming revision has a common ancestor with the active revision, we can use this to look up any attachments with revpos earlier (or matching) the common ancestor.