JWT validation and authentication, session support for OIDC auth
When OIDC is enabled in the database config, SG supports authentication using a bearer ID token.

Also adds support for SG session creation by the `_oidc_callback` endpoint.  When the ID token obtained during callback processing is successfully validated, it will create a Sync Gateway session and set the cookie on the response.  Session creation can be disabled using the disable_session property in the OIDC config.

When register=true in the OIDC config, SG users will be autocreated based on the subject and (optional) email in the ID token claims.