Expiry support for SG documents (#1881)
* Initial implementation for TTL

* Expiry support for SG documents

When writing a document to Sync Gateway (via PUT doc or bulk_docs), users can set the '_exp' property in the body of the document.  When Sync Gateway writes the document, it will:
 - Set a Couchbase Server expiry on the document, based on the exp value (see below for format)
 - Strip the _exp property from the document body (to avoid compatibility issues around the leading underscore when that document is replicated elsewhere)
 - Set an expiry property - `exp` - in the document's sync metadata, to provide some visibility on the expiry value

When Sync Gateway reads a document (via document GET, bulk_get), it will recreate the _exp property in the body as an ISO-8601 only if the query string includes show_exp=true .

Supported formats for the incoming _exp value:
 - JSON number.  Will be treated as a Couchbase Server expiry value (ttl in second when less than 30 days, unix time when greater)
 - JSON string (numeric format) - same as JSON number
 - JSON string (as ISO-8601 date).  Converted to Couchbase server expiry value.
 - JSON null.  Sets expiry to zero (no expiry)

Note that subsequent updates to the document will use the expiry value on the update.  If no expiry value is set on a new revision of a document that was previously set to expire, the new version of the document will have no expiry set.

Fixes #1823

* Updates based on code review feedback