Do not require callback_url, validation_key for OIDC authentication (#1889)
* Do not require callback_url, validation_key for OIDC authentication

If callback_url isn't present, SG builds the callback based on the request scope.  This wasn't working for ID token-based auth.  While fixing, realized that the existing callback URL initialization code wasn't synchronized, and also wouldn't work if the client was initialized before the callback url was set on the provider.  Modified to use a callback function to calculate the callback URL, and enforced that this function be passed as a parameter provider.GetClient(), and only called once during client initialization.  To support cases where the db isn't set on the request (ID token auth flow), switched to use PathVar("db") directly instead of getting it from h.rq.db.  It requires some passing around of the callback function (from the handler, to auth, eventually to GetClient(), but this was the cleanest approach without completely rebuilding the lazy initialization of the client.  Fixes #1883

Also removed requirement for Validation Key - clients using the implicit flow don't need to define a validation key.  Added a warning on startup if the validation key isn't defined to make it clear that the auth code flow won't work if the validation key isn't present.  Fixes #1877

* Code style fix based on PR review